# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted programming learning platform",
        "title": "CourseForge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/courses": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "List all courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.CourseItem"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{courseSlug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Get full course tree",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CourseDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{courseSlug}/tracks/{trackSlug}/topics/{topicSlug}/units/{unitSlug}/tasks/{taskSlug}/statement": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get task statement (markdown)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Track slug",
                        "name": "trackSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Topic slug",
                        "name": "topicSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unit slug",
                        "name": "unitSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task slug",
                        "name": "taskSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{courseSlug}/tracks/{trackSlug}/topics/{topicSlug}/units/{unitSlug}/tasks/{taskSlug}/template": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get task code template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Track slug",
                        "name": "trackSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Topic slug",
                        "name": "topicSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unit slug",
                        "name": "unitSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task slug",
                        "name": "taskSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language (e.g. go, python)",
                        "name": "lang",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{courseSlug}/tracks/{trackSlug}/topics/{topicSlug}/units/{unitSlug}/tasks/{taskSlug}/tests": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get task test file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Track slug",
                        "name": "trackSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Topic slug",
                        "name": "topicSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unit slug",
                        "name": "unitSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task slug",
                        "name": "taskSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language (e.g. go, python)",
                        "name": "lang",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/courses/{courseSlug}/tracks/{trackSlug}/topics/{topicSlug}/units/{unitSlug}/theory": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Get theory file (markdown)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Track slug",
                        "name": "trackSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Topic slug",
                        "name": "topicSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unit slug",
                        "name": "unitSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/progress/{courseSlug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get course progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progress.Progress"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/progress/{courseSlug}/tasks/{taskSlug}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Mark task done or undone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course slug",
                        "name": "courseSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task slug",
                        "name": "taskSlug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProgressUpdate"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/run": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run"
                ],
                "summary": "Execute code",
                "parameters": [
                    {
                        "description": "Run request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RunReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.RunResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/runners": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runners"
                ],
                "summary": "List registered language drivers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/runner.LangDriver"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "runners"
                ],
                "summary": "Add a new language driver",
                "parameters": [
                    {
                        "description": "Driver config",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddDriverReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/runners/{lang}": {
            "delete": {
                "tags": [
                    "runners"
                ],
                "summary": "Delete a language driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language key (e.g. go)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "runners"
                ],
                "summary": "Partially update an existing language driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language key (e.g. go)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PatchDriverReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AddDriverReq": {
            "type": "object",
            "properties": {
                "driver": {
                    "$ref": "#/definitions/runner.LangDriver"
                },
                "lang": {
                    "type": "string",
                    "example": "rust"
                }
            }
        },
        "api.CourseDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TrackItem"
                    }
                }
            }
        },
        "api.CourseItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.PatchDriverReq": {
            "type": "object",
            "properties": {
                "ext": {
                    "type": "string"
                },
                "run_cmd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "test_cmd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "test_ext": {
                    "type": "string"
                }
            }
        },
        "api.ProgressUpdate": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "api.RunReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "package main\nfunc main() {}"
                },
                "language": {
                    "type": "string",
                    "example": "go"
                },
                "test_code": {
                    "type": "string"
                },
                "timeout_sec": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "api.RunResp": {
            "type": "object",
            "properties": {
                "duration_ms": {
                    "type": "integer"
                },
                "exit_code": {
                    "type": "integer"
                },
                "stderr": {
                    "type": "string"
                },
                "stdout": {
                    "type": "string"
                },
                "timed_out": {
                    "type": "boolean"
                }
            }
        },
        "api.TaskItem": {
            "type": "object",
            "properties": {
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.TopicItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UnitItem"
                    }
                }
            }
        },
        "api.TrackItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TopicItem"
                    }
                }
            }
        },
        "api.UnitItem": {
            "type": "object",
            "properties": {
                "has_theory": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TaskItem"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "progress.Progress": {
            "type": "object",
            "properties": {
                "completed_tasks": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "course_slug": {
                    "type": "string"
                }
            }
        },
        "runner.LangDriver": {
            "type": "object",
            "properties": {
                "ext": {
                    "description": "source file extension, e.g. \".go\"",
                    "type": "string"
                },
                "run_cmd": {
                    "description": "e.g. [\"go\", \"run\", \"{file}\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "test_cmd": {
                    "description": "e.g. [\"go\", \"test\", \"{file}\", \"{testfile}\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "test_ext": {
                    "description": "test file name suffix, e.g. \"_test.go\"",
                    "type": "string"
                }
            }
        }
    }
}